backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.2
aiosqlite==0.22.1
greenlet==3.3.2
